fix: move PostModel
