+++ project/apps/blog/src/app/posts/posts.module.ts
@@ -4,7 +4,7 @@ import { PostsController } from './posts.controller';
 import { PostRepository } from './repository/post.repository';
 import { PostsContentModule } from './post-content/post-content.module';
 import { MongooseModule } from '@nestjs/mongoose';
-import { PostModel, PostSchema } from './posts.model';
+import { PostModel, PostSchema } from './models/posts.model';
 
 @Module({
   imports: [
